se añadió la estructura inicial para trabajar con el flujo de creación quiz
