+++ app/(crearQuiz)/creacionQuiz.tsx
@@ -126,6 +126,8 @@ export default function CrearQuiz() {
                         createdat: new Date().toISOString(),
                     };
                     console.log(objQuiz);
+                    router.push({ pathname: 'confirmarQuiz', params: { jsonExercise: JSON.stringify(objQuiz) } });
+                    /*
                     ExerciseService.createExercise(objQuiz).then((response) => {
                       if (response.error) {
                         Alert.alert(
@@ -135,7 +137,7 @@ export default function CrearQuiz() {
                       } else {
                         router.replace("finalizarQuiz");
                       }
-                    });
+                    });*/
                     
                 }}
             >
