+++ services/exercise.tsx
@@ -76,7 +76,6 @@ export class ExerciseService {
     exercise: Tables<"exercises"> | null;
     error: PostgrestError | null;
   }> {
-    console.log("Entra a servicio de get exercis");
     const { data, error } = await supabase
       .from("exercises") // Cambia 'exercises' al nombre de tu tabla si es diferente
       .select("*")
